+++ src/api.js
@@ -178,6 +178,8 @@ class API {
       lang,
       redirectURI = this.config.redirectURI,
       providers,
+      theme,
+      providersPerLine,
     } = options;
     const { clientId, connectURL } = this.config;
 
@@ -193,6 +195,8 @@ class API {
       demo: this.config.demo ? true : undefined,
       // 'provider' param in priority, skip 'providers' param if 'provider' is set
       providers: !provider && Array.isArray(providers) && providers.length ? providers.join(',') : undefined,
+      theme: ['light', 'dark'].includes(theme) ? theme : 'light',
+      providers_per_line: (providersPerLine && ['1', '2'].includes(providersPerLine.toString())) ? providersPerLine.toString() : '2',
     };
 
     // Cleanup blank params
@@ -255,10 +259,10 @@ class API {
       try {
         this._widgetOpened = true;
         const {
-          provider, providers, accountId, lang,
+          provider, providers, accountId, lang, theme, providersPerLine,
         } = options;
         const { url, token } = await this.getConnectData({
-          provider, providers, accountId, lang,
+          provider, providers, accountId, lang, theme, providersPerLine,
         });
 
         this.iframe = appendVezgoIframe();