+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the Vezgo SDK across versions will be documented in this 
 
 ## [Unreleased]
 
+## [1.0.3]
+- Add `theme` (light | dark) support for `user.connect()`, `user.reconnect()` & `user.getConnectUrl()`.
+- Add `providersPerLine` (1 | 2) support for `user.connect()`, `user.reconnect()` & `user.getConnectUrl()`.
+
 ## [1.0.2]
 - Add keywords
 - Add annotation
@@ -107,8 +111,9 @@ Note: Latest Vezgo Connect requires to be opened via `POST` request with `token`
 
 - Add browser auth parameters (`authEndpoint`, `auth { params, headers }`, `authorizer`) to `Vezgo.init()`
 
-[Unreleased]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.2...HEAD
-[1.0.1]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.1...v1.0.2
+[Unreleased]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.3...HEAD
+[1.0.3]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.2...v1.0.3
+[1.0.2]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/wealthica/vezgo-sdk-js/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/wealthica/vezgo-sdk-js/compare/v0.0.16...v1.0.0
 [0.0.16]: https://github.com/wealthica/vezgo-sdk-js/compare/v0.0.15...v0.0.16